moved PoseExtractor to separate file and other formatting changes
